requirements.txt
asgiref==3.3.1
certifi==2019.9.11
Django==3.1.3
django-rest-framework==0.1.0
django-tables2==2.3.3
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
pipenv==2018.11.26
pytz==2020.4
sqlparse==0.4.1
virtualenv==16.7.7
virtualenv-clone==0.5.3
